02_Repository/ejemplo_practico/requirements.txt
fastapi==0.113.0
SQLAlchemy==2.0.34
uvicorn==0.30.6
psycopg2==2.9.9
fastapi-utils==0.7.0
pydantic==2.9.0
mysql-connector-python==9.0.0